returned value can now be prefixed/suffixed with a given string
If the returned value is of string type, then the sixth parameter is taken under consideration.
If a simple string is passed, then the returned value will be prefixed with that.

In order to use and/or a suffix as well, then the string has to be in special format: 'PRE:STRING\,SUF:STRING'

In general all of those values are accepted:

'PRE:STRING\,SUF:STRING'
'SUF:STRING\,PRE:STRING'
'PRE:STRING'
'SUF:STRING'

But those values will return the whole string:

'STRING\,SUF:STRING'
'SUF:STRING\,STRING'